dom/workers/test/serviceworkers/test_update_missing_imported_script.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1290951 - Test update after a new verion of mainscipt that doesn't need an imported script anymore.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
<!--
  If the principal is not set, accessing self.caches in the worker will crash.
-->
</head>
<body>
<p id="display"></p>
<div id="content"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  // We have the iframe register for its own scope so that this page is not
  // holding any references when we GC.
  function register() {
    return Promise.resolve()
      .then(_ => navigator.serviceWorker.register("http://mochi.test:8888/" +
                                                  "tests/dom/workers/test/" +
                                                  "serviceworkers/" +
                                                  "bug1290951_worker_main.sjs"))
      .then(r => {
        ok(r, "Should be a registration.");

        // Get the newest worker.
        var worker = r.installing || r.waiting || r.active;

        // Wait for the worker becoming active.
        return new Promise(aResolve => {
          if (worker.state === "activated") {
            aResolve();
            return;
          }

          worker.addEventListener("statechange", _ => {
            if (worker.state === "activated") {
              aResolve();
            }
          });
        });
      });
  }

  function update() {
    return Promise.resolve()
      .then(_ => navigator.serviceWorker.getRegistration())
      .then(r => {
        return new Promise(aResolve => {
          r.addEventListener("updatefound", aResolve);
          r.update();
        });
      });
  }

  function unregister() {
    return Promise.resolve()
      .then(_ => navigator.serviceWorker.getRegistration())
      .then(r => r.unregister())
  }

  function runTest() {
    register()
      .then(register)
      .then(update)
      .then(unregister)

      .catch(e => ok(false, "Some test failed with error " + e))
      .then(SimpleTest.finish);
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
